requirements.txt
pandas==1.3.3
pysftp==0.2.9
python-dotenv==1.0.1
requests==2.27.1
cx-Oracle==8.3.0
openpyxl==3.1.3
xlsxwriter==3.2.2
mysql-connector-python==8.0.32
tqdm==4.67.1